- added lots of stuff - skills - shop - new HUD - effects - menu window - other stuff

+++ src/game/entities/WizardBullet.java
@@ -10,11 +10,11 @@ import org.newdawn.slick.geom.Point;
 
 public class WizardBullet extends Bullet{
 
-	public WizardBullet(Point p, int vx, int vy, GOType type) throws SlickException {
+	public WizardBullet(Point p, float vx, float vy, GOType type) throws SlickException {
 		super(p, vx, vy, type);
 
 		// init sprite
-		sprite = new Image("res/wizard/normalatk.png").getSubImage(++vx*Game.TS, ++vy*Game.TS, Game.TS, Game.TS);
+		sprite = new Image("res/wizard/normalatk.png").getSubImage((int) ++vx *Game.TS, (int) ++vy * Game.TS, Game.TS, Game.TS);
 		// init bounds
 		
 		bounds.setHeight(sprite.getHeight());
